+++ fedcloud_monitoring_tools/data/vos.yaml
